test restructure; claning
